Don't clear a Unit's affiliation as it leaves an Army
It was causing the bug from the previous commit.  This should be okay because there's
no cross-dependency and it's fine if a unit has an affiliation but the Army it's in doesn't have it as a child, because the only time that
should happen is during a combat scene, where it will be re-added later, and as the unit
is being deleted anyway.